+++ R/center.R
@@ -225,7 +225,6 @@ center.data.frame <- function(x,
 }
 
 
-
 #' @export
 center.grouped_df <- function(x,
                               select = NULL,
@@ -275,4 +274,32 @@ center.grouped_df <- function(x,
   # set back class, so data frame still works with dplyr
   attributes(args$x) <- args$info
   args$x
+}
+
+
+# methods -------------------------
+
+#' @export
+print.dw_transformer <- function(x, ...) {
+  print(as.vector(x), ...)
+  vector_info <- NULL
+  if (!is.null(attributes(x)$scale)) {
+    # attributes for center() / standardize()
+    vector_info <- sprintf(
+      "(center: %.2g, scale = %.2g)\n",
+      attributes(x)$center,
+      attributes(x)$scale
+    )
+  } else if (!is.null(attributes(x)$range_difference)) {
+    # attributes for normalize() / rescale()
+    vector_info <- sprintf(
+      "(original range = %.2g to %.2g)\n",
+      attributes(x)$min_value,
+      attributes(x)$min_value + attributes(x)$range_difference
+    )
+  }
+  if (!is.null(vector_info)) {
+    insight::print_color(vector_info, color = "grey")
+  }
+  invisible(x)
 }